Change address format to use 33-byte keys
Instead of 32-byte x-only keys, use compressed public keys
for the silent payments address.

This is to avoid parity mismatches, especially in the case of adding
labels to the spend public key, which will cause the sign of the
y point to flip between even and odd.

Without the extra byte, we would need the sender and receiver to
constantly negate the public and private keys at each step, which
adds complexity and computational expense.