Replaces hash for sha256, and adds first iter for consistency
`sha256` is used as the hashing function up until "Using all inputs", from
where `hash` is used. The same naming should be used for consistency.
Furthermore, the example for all inputs does not include the step pace
for `outpoint_hash`, but it does for `a`. Given they both use the same
pace and number of iterations, it'll be best to make them look the same